Feb 24, 2025; I am working on adding the map, and adding the menu. There are still some crucial errors though.

+++ app/src/main/java/com/example/earthquakelistproject/Geometry.kt
@@ -5,5 +5,5 @@ import kotlinx.parcelize.Parcelize
 // it should be parcelize, along with some other classes
 @Parcelize
 data class Geometry(
-    var coordinates: List<Int>
+    var coordinates: List<Double>
 ):Parcelable
